requirements.txt
pyspark
delta-spark
requests
pandas
